make denominator and nominator in split info f64 (was u64)

+++ src/quotes.rs
@@ -226,12 +226,12 @@ pub struct Split {
     /// wherever you had one before the split. (Here the numerator is 1 and
     /// denom is 5). A reverse split is considered as nothing but a regular
     /// split with a numerator > denom.
-    pub numerator: u64,
+    pub numerator: f64,
     /// Denominator of the split. For instance a 1:5 split means you get 5 share
     /// wherever you had one before the split. (Here the numerator is 1 and
     /// denom is 5). A reverse split is considered as nothing but a regular
     /// split with a numerator > denom.
-    pub denominator: u64,
+    pub denominator: f64,
     /// A textual representation of the split.
     #[serde(rename = "splitRatio")]
     pub split_ratio: String,
